webapp/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref>=3.8.0
certifi==2025.6.15
charset-normalizer>=3.4.0
click>=8.2.0
fastapi>=0.115.0
greenlet>=3.2.0
h11>=0.16.0
httptools>=0.6.0
idna>=3.10
Jinja2>=3.1.6
ldap3>=2.9.1
markdown-it-py==3.0.0
MarkupSafe>=3.0.0
mdurl==0.1.2
psutil>=7.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pydantic>=2.11.0
pydantic_core>=2.33.0
Pygments==2.19.1
PyMySQL>=1.1.1
python-dateutil>=2.9.0
python-dotenv==1.1.0
python-ldap>=3.4.4
python-multipart>=0.0.20
python-on-whales>=0.77.0
PyYAML==6.0.2
requests>=2.32.0
rich==14.0.0
shellingham==1.5.4
six>=1.17.0
sniffio>=1.3.0
SQLAlchemy>=2.0.0
starlette>=0.46.0
tqdm>=4.67.0
typer>=0.16.0
uvicorn[standard]>=0.34.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
alembic>=1.13.0
